mysite/requirements.txt
asgiref
DateTime
Django
gunicorn
packaging
pytz
sqlparse
zope.interface
--index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments